feat(utils): make package public
