+++ packages/utils/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@angular-eslint/utils",
   "version": "12.3.1",
-  "private": true,
   "license": "MIT",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
