+++ packages/utils/src/index.ts
@@ -1,2 +1,2 @@
 export * from './rules-tester';
-export * from './test-helpers';
+export * from './convert-annotated-source-to-failure-case';
